+++ extensions/vscode-aml/src/web/aml.ts
@@ -1,11 +1,18 @@
 import {Database, ParserResult} from "@azimutt/models";
+// @ts-ignore
+import {AmlAst} from "@azimutt/aml/out/amlAst";
 
 // dynamic import to use ESM libs in a CommonJS package (required by VS Code extension)
 const amlLib = import("@azimutt/aml");
+const amlParser = import("@azimutt/aml/out/amlParser.js");
 const sqlLib = import("@azimutt/parser-sql");
 
 export async function parseAml(input: string): Promise<ParserResult<Database>> {
     return (await amlLib).parseAml(input)
+}
+
+export async function parseAmlAst(input: string): Promise<ParserResult<AmlAst>> {
+    return (await amlParser).parseAmlAst(input, {})
 }
 
 export async function generateAml(db: Database): Promise<string> {